Initial version of a PIM events plugin for Plasma Clock Applet
This implements a plugin through ETMCalendar. It eats a lot of
memory. The UI in Plasma won't be available until Plasma 5.7,
so there's no hurry in getting this in (I'm targetting 16.08).

There is no configuration yet, because the current code in
Plasma Digital Clock applet does not implement it yet.