MessageViewer plugin: create EventEdit widget on demand.
Summary:
This avoids creating a CollectionComboBox on startup (which includes a Monitor,
an ETM and 3 proxies on top), which might never be used if the user doesn't
use the "create event" feature.

The same change should of course be done to noteedit and todoedit.

The unittest had to be adjusted because it was checking that the widget
was created upfront and hidden, which is exactly what this patch changes.

Test Plan:
Ctrl+E in kmail still brings up the event edit widget,
without any noticeable delay the first time.

Reviewers: knauss, mlaurent

Reviewed By: knauss, mlaurent

Subscribers: knauss, kde-pim, #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D2161