Re-enable etm_usage, port from CR to Monitor
Strictly speaking this change is not even necessary, the build could've
been easilly fixed simply by forward-declaring or including CR header
in our headers (and not relaying on ETM header to do it). But since
all client apps should only use Monitors and not CRs, let's port the
examples, so they are closer to what we want.