PimEventsPlugin: switch from ETMCalendar to custom EventModel
It is hard to tame the ETMCalendar to list only the right events
and listen only to the right changes while not keeping everything
loaded in the memory all the time.

The EventModel is a lightweight bridge between Akonadi and the
CalendarBase. It inserts and removes data from the CalendarBase
based on currently enabled collections and handles updates.
This solves various issues with populating the model initially,
listening to the correct changes and it also reduces memory usage
by not keeping events from disabled calendars in memory.

BUG: 369606
BUG: 368832
FIXED-IN: 5.3.3