Introduce GnuPG WKS and PGP keys formatter plugins
This adds two new body part formatter plugins for messageviewer:
one for application/pgp-keys and one for application/vnd.gnupg.wks
mimetypes.

The application/pgp-keys formatter simply prints a message saying
that this part is an OpenPGP key and explains what a key is, and has
buttons to display key details and to import the key to user's
keychain or open the key in Kleopatra.

The application/vnd.gnupg.wks formatter handles WKS Publishing
Comfirmation request messages. Those are automated emails sent
by WKS server when user request key publishing. The formatter
explains what this email is, why it was received and has a
'Register the key' button which, when clicked, generates a
response email for the server and deletes the email. The response
is also deleted.

Differential Revision: https://phabricator.kde.org/D3140