Try harder to parse localized three letter month abbreviations
For en_US the data in QLocale matches commonly used three letter month
shortcuts, for other languages the strings in QLocale can be longer. If we
fail to parse a date containing the 'MMM' format, we'll not try to extend
the contained three letter month codes to what QLocale expects.