Rework data extraction
We now no longer stop after the first found element, or unconditionally
prefer structured over unstructured data, but merge everything together
we find in the entire email.

This helps in cases of e.g. structured boarding pass data and a pkpass
file in the same email. Even in the best case of the structured data being
a superset of the data found in the pkpass file (which is a purely
theoretical scenario), we'd still get the association with the pkpass file
added by this which helps the mobile app to find the right pass.