Fix GrantleeTheme::ThemeManager being leaked.
Summary:
When a Viewer was deleted, its ThemeManager was kept alive.
So it would still react to KDirWatch (when installing files) and it
would try to update a menu (QActionGroup) that was deleted with the
viewer.

BUG: 383070
FIXED-IN: 18.04

Test Plan:
KMail, Enter to open separate viewer, Alt+F4, (uncommitted)
debug output shows that now the ThemeManager is deleted, while it wasn't before.

Reviewers: mlaurent

Reviewed By: mlaurent

Subscribers: #kde_pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D12431