Fix send to KDE Connect for older KDE Connect version
Summary:
When I wrote the code I assumed that by the time it is released the corresponding KDE Connect changes would be released, too. History proved me wrong. The openFile method will only be available with KDE Connect 1.4 whose release
is not yet scheduled.

For older versions of KDE Connect shareUrl is used now. That transfers the file but does not open it.

Test Plan: Tested with KDE Connect 1.3.3 and master (with version bumped to 1.4).

Reviewers: vkrause, #kde_pim_kmail

Reviewed By: vkrause

Subscribers: kde-pim

Tags: #kde_pim, #kde_pim_kmail, #kde_itinerary

Differential Revision: https://phabricator.kde.org/D17822