Calendar formatter: port to PimUniqueApplication::activateApplication
Summary:
The whole "if korganizer or kontact is running, bring it to the front, otherwise
start korganizer" logic becomes a single line of code.
I'm not 100% sure about the Windows code using QProcess; clearly it
relies on DBus anyway so the new way should work, but it mentions
korganizer-mobile.sh (which could be called by the DBus service file then...)

Test Plan:
Click "Check My Calendar" in an invitation email.
In KMail, it starts standalone korganizer.
In Kontact, it switches to the calendar plugin.

Reviewers: vkrause, dvratil, mlaurent

Reviewed By: vkrause

Subscribers: kde-pim

Tags: #kde_pim

Differential Revision: https://phabricator.kde.org/D22622