Reenable lunarphases korganizer plugin, use icon instead of text
Also remove Lunarphases::createWeekElements(), old code would add an icon
for the whole week if on start of the week there is a lunar event, which
is not really something wanted.

sc-status-moon-phase-full.svg is a copy of Breeze's
weather-clear-night.svg of KF 5.84, with the other icons derived from it.

BUG: 30331
FIXED-IN: 21.12