Don't require calendar invite to be inside a message/rfc822 message
When the calendar invitation is inside an encrypted and signed parts, the
top-level node is actually not a message, but usually the temporary node
created during decryption or signature verification.

The fact the the top-level node must be an actual message was only needed in order
to extract the sender. We preserve this, but don't require it - in many cases we
don't actually need to know the sender in the IcalFormatter.