Remove Akonadi Relation support from Create Event/Todo plugins
The relation support is going to be removed from Akonadi. It was only
ever supported by the Kolab Resource, cross-resource relations were
impossible to be sync and we had no UI to interact with the relations.
We can re-introduce this through search index later on, if we want to
provide information about related entities.